ostad-shopify-app-assignment-1/requirements/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements/requirements.in
#
asgiref==3.7.2
    # via
    #   django
    #   django-browser-reload
certifi==2023.11.17
    # via sentry-sdk
django==4.2.7
    # via
    #   -r requirements/requirements.in
    #   django-browser-reload
    #   django-shopify-auth
    #   django-tailwind
django-browser-reload==1.12.1
    # via django-tailwind
django-environ==0.10.0
    # via -r requirements/requirements.in
django-shopify-auth @ git+https://github.com/farhanmasud/django-shopify-auth.git
    # via -r requirements/requirements.in
django-tailwind==3.5.0
    # via -r requirements/requirements.in
ecdsa==0.18.0
    # via python-jose
pillow==9.5.0
    # via -r requirements/requirements.in
psycopg2-binary==2.9.6
    # via -r requirements/requirements.in
pyactiveresource==2.2.2
    # via shopifyapi
pyasn1==0.5.1
    # via
    #   python-jose
    #   rsa
pyjwt==2.8.0
    # via shopifyapi
python-jose==3.3.0
    # via django-shopify-auth
pyyaml==6.0.1
    # via shopifyapi
rsa==4.9
    # via python-jose
sentry-sdk==1.18.0
    # via -r requirements/requirements.in
shopifyapi==12.3.0
    # via
    #   -r requirements/requirements.in
    #   django-shopify-auth
six==1.16.0
    # via
    #   ecdsa
    #   pyactiveresource
    #   shopifyapi
sqlparse==0.4.4
    # via django
typing-extensions==4.8.0
    # via asgiref
ua-parser==0.18.0
    # via -r requirements/requirements.in
urllib3==2.1.0
    # via sentry-sdk
whitenoise==6.4.0
    # via -r requirements/requirements.in

# The following packages are considered to be unsafe in a requirements file:
# setuptools
